requirements.txt
Django>=5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
gunicorn>=20.1
djangorestframework-simplejwt
drf-spectacular
djangorestframework